Reports: Attempt to speed up the reports UI
Needs new commands in crmsh to really work.

The problem now is that every crmsh history command
gets slower and slower and slower the larger the
logfile to analyse is.

What I'd want to do is call crmsh history once and
get it to return as much information about transitions
as possible.